+++ common-client/src/main/java/de/maxhenkel/voicechat/gui/volume/AdjustVolumeList.java
@@ -53,7 +53,7 @@ public class AdjustVolumeList extends ListScreenListBase<VolumeEntry> {
     }
 
     private void addOfflinePlayers(Collection<PlayerState> onlinePlayers) {
-        for (UUID uuid : VoicechatClient.VOLUME_CONFIG.getPlayerVolumes().keySet()) {
+        for (UUID uuid : VoicechatClient.PLAYER_VOLUME_CONFIG.getVolumes().keySet()) {
             if (uuid.equals(Util.NIL_UUID)) {
                 continue;
             }