+++ fabric/changelog.md
@@ -10,3 +10,4 @@
 - Fixed French translation
 - Fixed speaker button alignment
 - Added audio level meters to adjust volumes section
+- Reworked player volume adjustment